Quickly find MB quarter sections on a map

Plot the output of search_legal() or search_coord() or similar dataframe onto an interactive Leaflet map

Usage

map_quarter(x, map.type = "Esri.WorldImagery")

Arguments

x

Output from search_legal() or search_coord() or similar dataframe

map.type

Base map: see https://leaflet-extras.github.io/leaflet-providers/preview/ for available options

Value

A Leaflet map

Details

Data frame must include the following three column names:

legal - A character that specifies the legal land description (NE-11-33-29W1)

long - A numeric specifying the longitude in decimal degrees (-97.6)

lat - A numeric specifying the latitude in decimal degrees (49.1)

Examples

search1 <- search_legal(x = c("NE-11-33-29W1", "SW-20-2-1W1"))
#> Data doesn't exist, please download with `quarters_dl()` first
map_quarter(x = search1)
#> Error in map_quarter(x = search1): Must provide a dataframe
search2 <- search_coord(long = c(-101.4656, -99.99768), lat = c(51.81913, 49.928926))
#> Error in search_coord(long = c(-101.4656, -99.99768), lat = c(51.81913,     49.928926)): Data doesn't exist, please download with `quarters_dl()` first
map_quarter(x = search2)
#> Error in eval(expr, envir, enclos): object 'search2' not found