Finding Manitoba, Canada, quarter sections made easier

mbquartR is an R package for locating quarter sections in Manitoba using lat/long coordinates or the legal land descriptions using the Manitoba Original Survey Legal Descriptions data set from the DataMB website (https://geoportal.gov.mb.ca/).

Details

There are four main functions in this package:

Download the Manitoba Original Survey Legal Descriptions data set

quarters_dl()

Search for legal land descriptions using lat and long coordinates

search_coord()

Search for lat and long coordinates using legal land descriptions

search_legal()

Plot search results on an interactive map

map_quarter()

See also

Author

Maintainer: Alex Koiter koitera@brandonu.ca (ORCID) (Associate Professor, Department of Geography and Environment, Brandon University <https://alexkoiter.ca/>)

Other contributors:

Emily H Markowitz (Research Fisheries Biologist, Alaska Fisheries Science Center NOAA Fisheries | U.S. Department of Commerce. <https://emilyhmarkowitz.github.io/emilyhmarkowitz> Emily reviewed the package for rOpenSci, see <https://github.com/ropensci/software-review/issues/658>) [reviewer]

Sheila M Saia (Environmental Data Scientist, Tetra Tech. <https://sheilasaia.rbind.io/> Sheila reviewed the package for rOpenSci, see <https://github.com/ropensci/software-review/issues/658>) [reviewer]